Redirect users through the login flow on 401
Note: I have elected not to write tests for this because changing
window.location during a test run would bork up the tests.
