+++ src/fetch.js
@@ -1,5 +1,8 @@
 import { API_ROOT } from './constants';
 import * as isomorphicFetch from 'isomorphic-fetch';
+import store from '~/store';
+import checkLogin from '~/session';
+import { setToken } from '~/actions/authentication';
 
 /*
  * Sinon cannot stub out a function in a function-only module.
@@ -7,6 +10,12 @@ import * as isomorphicFetch from 'isomorphic-fetch';
  */
 export function rawFetch(...args) {
   return isomorphicFetch['default'](...args);
+}
+
+function expireSession() {
+  const next = { location: window.location };
+  store.dispatch(setToken(null, null, null, null));
+  checkLogin(next);
 }
 
 export function fetch(token, _path, _options) {
@@ -34,6 +43,9 @@ export function fetch(token, _path, _options) {
       // eslint-disable-next-line no-param-reassign
       response.statusCode = status;
       if (status >= 400) {
+        if (status === 401) {
+          expireSession();
+        }
         reject(response);
       } else {
         accept(response);
